Crate libcrux_traits

Re-exports

pub use libcrux_secrets;

Modules

aead

digest

This module provides common interface traits for digest/hashing functionality.

ecdh

This module provides a common interface trait for Elliptic Curve Diffie-Hellman (ECDH) key exchange.

kem

This module provides a common interface trait for key encapsulation mechanisms (KEMs).

Macros

impl_aead_slice_trait

Implements slice::Aead based on arrayref::Aead.

impl_aead_typed_owned

impl_digest_incremental_trait

Implements DigestIncremental for any arrayref::DigestIncremental.

impl_ecdh_slice_trait

Implements EcdhSlice for any $ty : arrayref::EcdhArrayref with the given array bounds.

impl_hash_trait

Implements Hash for any arrayref::Hash.

impl_trait

Implements Kem for any arrayref::Kem

Traits

Digest

A Hash algorithm returning hashes of length HASH_LEN.