libcrux_traits::aead::owned

Trait Aead

pub trait Aead<const KEY_LEN: usize, const TAG_LEN: usize, const NONCE_LEN: usize> {
    // Required methods
    fn keygen(rand: &[U8; KEY_LEN]) -> Result<[U8; KEY_LEN], KeyGenError>;
    fn encrypt<const MSG_LEN: usize>(
        key: &[U8; KEY_LEN],
        nonce: &[U8; NONCE_LEN],
        aad: &[u8],
        plaintext: &[U8; MSG_LEN],
    ) -> Result<([u8; MSG_LEN], [U8; TAG_LEN]), EncryptError>;
    fn decrypt<const MSG_LEN: usize>(
        key: &[U8; KEY_LEN],
        nonce: &[U8; NONCE_LEN],
        aad: &[u8],
        ciphertext: &[u8; MSG_LEN],
        tag: &[U8; TAG_LEN],
    ) -> Result<[U8; MSG_LEN], DecryptError>;
}

An Authenticated Encryption with Associated Data (AEAD) scheme

Some implementors of this trait may impose stronger restrictions on the inputs than described here. Check the documentation of the types implementing this trait to make sure which inputs are valid.

Required Methods

fn keygen(rand: &[U8; KEY_LEN]) -> Result<[U8; KEY_LEN], KeyGenError>

Generate a new key. Consumes the entire randomnes.

fn encrypt<const MSG_LEN: usize>( key: &[U8; KEY_LEN], nonce: &[U8; NONCE_LEN], aad: &[u8], plaintext: &[U8; MSG_LEN], ) -> Result<([u8; MSG_LEN], [U8; TAG_LEN]), EncryptError>

Encrypt a plaintext message, producing a ciphertext and an authentication tag.

fn decrypt<const MSG_LEN: usize>( key: &[U8; KEY_LEN], nonce: &[U8; NONCE_LEN], aad: &[u8], ciphertext: &[u8; MSG_LEN], tag: &[U8; TAG_LEN], ) -> Result<[U8; MSG_LEN], DecryptError>

Decrypt a ciphertext, verifying its authenticity.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.