libcrux_traits::aead::slice

Trait Aead

pub trait Aead {
    // Required methods
    fn keygen(key: &mut [U8], rand: &[U8]) -> Result<(), KeyGenError>;
    fn encrypt(
        ciphertext: &mut [u8],
        tag: &mut [U8],
        key: &[U8],
        nonce: &[U8],
        aad: &[u8],
        plaintext: &[U8],
    ) -> Result<(), EncryptError>;
    fn decrypt(
        plaintext: &mut [U8],
        key: &[U8],
        nonce: &[U8],
        aad: &[u8],
        ciphertext: &[u8],
        tag: &[U8],
    ) -> Result<(), DecryptError>;
}

An Authenticated Encryption with Associated Data (AEAD) scheme.

Some implementors of this trait may impose stronger restrictions on the inputs than described here. Check the documentation of the types implementing this trait to make sure which inputs are valid.

Required Methods

fn keygen(key: &mut [U8], rand: &[U8]) -> Result<(), KeyGenError>

Generate a new key. Consumes the entire randomnes. It is the responsibility of the caller to ensure rand is random and long enough (usually as long as key).

fn encrypt( ciphertext: &mut [u8], tag: &mut [U8], key: &[U8], nonce: &[U8], aad: &[u8], plaintext: &[U8], ) -> Result<(), EncryptError>

Encrypt a plaintext message, producing a ciphertext and an authentication tag. The arguments plaintext and ciphertext must have the same length.

fn decrypt( plaintext: &mut [U8], key: &[U8], nonce: &[U8], aad: &[u8], ciphertext: &[u8], tag: &[U8], ) -> Result<(), DecryptError>

Decrypt a ciphertext, verifying its authenticity. The arguments plaintext and ciphertext must have the same length.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.