libcrux_traits::aead

Module typed_refs

This module contains the trait and related errors for an Authenticated Encryption with Associated Data (AEAD) scheme that takes array references as arguments and writes outputs to mutable array references.

Structs

KeyMut

A Key with the given algorithm. The bytes are borrowed mutably. Contains a marker for which algorithm the key is to be used with.

KeyRef

A Key with the given algorithm. The bytes are borrowed. Contains a marker for which algorithm the key is to be used with.

NonceRef

A nonce with the given algorithm. The bytes are borrowed. Contains a marker for which algorithm the key is to be used with.

TagMut

A mutable tag with the given algorithm. The bytes are borrowed mutably. Contains a marker for which algorithm the key is to be used with.

TagRef

A tag with the given algorithm. The bytes are borrowed. Contains a marker for which algorithm the key is to be used with.

WrongLengthError

Enums

DecryptError

Error that can occur during decryption.

EncryptError

Error that can occur during encryption.

Traits

Aead

An Authenticated Encryption with Associated Data (AEAD) scheme. This trait is low-level and is mostly used for implementing other, more usable APIs.

Multiplexes

This trait is implemented by the multiplexing enum, once per actual algorithm. It allows users (and internal code) to convert between structs with the actual algorithm and the multiplexing enum.