libcrux_traits::aead::typed_refs

Struct TagMut

pub struct TagMut<'a, Algo> { /* private fields */ }

A mutable tag with the given algorithm. The bytes are borrowed mutably. Contains a marker for which algorithm the key is to be used with.

Implementations

impl<'a, Algo: Aead> TagMut<'a, Algo>

pub fn new_for_algo( algo: Algo, tag: &'a mut [U8], ) -> Result<Self, WrongLengthError>

Creates a new mutable tag for the provided algorithm. Checks that the length is correct.

impl<'a, Algo> TagMut<'a, Algo>

pub fn algo(&self) -> &Algo

Returns the algorithm this mutable tag should be used in