libcrux_traits::ecdh::owned

Trait EcdhOwned

pub trait EcdhOwned<const RAND_LEN: usize, const SECRET_LEN: usize, const PUBLIC_LEN: usize> {
    // Required methods
    fn generate_secret(
        rand: &[U8; RAND_LEN],
    ) -> Result<[U8; SECRET_LEN], GenerateSecretError>;
    fn secret_to_public(
        secret: &[U8; SECRET_LEN],
    ) -> Result<[u8; PUBLIC_LEN], SecretToPublicError>;
    fn derive_ecdh(
        public: &[u8; PUBLIC_LEN],
        secret: &[U8; SECRET_LEN],
    ) -> Result<[U8; PUBLIC_LEN], DeriveError>;
    fn validate_secret(
        secret: &[U8; SECRET_LEN],
    ) -> Result<(), ValidateSecretError>;

    // Provided method
    fn generate_pair(
        rand: &[U8; RAND_LEN],
    ) -> Result<([u8; PUBLIC_LEN], [U8; SECRET_LEN]), GenerateSecretError> { ... }
}

An Elliptic Curve Diffie-Hellman (ECDH) key exchange.

Required Methods

fn generate_secret( rand: &[U8; RAND_LEN], ) -> Result<[U8; SECRET_LEN], GenerateSecretError>

Generate a Diffie-Hellman secret value. It is the responsibility of the caller to ensure that the rand argument is actually random.

fn secret_to_public( secret: &[U8; SECRET_LEN], ) -> Result<[u8; PUBLIC_LEN], SecretToPublicError>

Derive a Diffie-Hellman public value from a secret value.

fn derive_ecdh( public: &[u8; PUBLIC_LEN], secret: &[U8; SECRET_LEN], ) -> Result<[U8; PUBLIC_LEN], DeriveError>

Derive a Diffie-Hellman shared secret from a public and a secret value.

This value is NOT (!) safe for use as a key and needs to be processed in a round of key derivation, to ensure both that the output is uniformly random and that unkown key share attacks can not happen.

fn validate_secret(secret: &[U8; SECRET_LEN]) -> Result<(), ValidateSecretError>

Check the validity of a Diffie-Hellman secret value.

Provided Methods

fn generate_pair( rand: &[U8; RAND_LEN], ) -> Result<([u8; PUBLIC_LEN], [U8; SECRET_LEN]), GenerateSecretError>

Generate a Diffie-Hellman secret value and derive the corresponding public value in one step.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.