libcrux_traits::kem::slice

Trait Kem

pub trait Kem {
    // Required methods
    fn keygen(
        ek: &mut [u8],
        dk: &mut [U8],
        rand: &[U8],
    ) -> Result<(), KeyGenError>;
    fn encaps(
        ct: &mut [u8],
        ss: &mut [U8],
        ek: &[u8],
        rand: &[U8],
    ) -> Result<(), EncapsError>;
    fn decaps(ss: &mut [U8], ct: &[u8], dk: &[U8]) -> Result<(), DecapsError>;
}

A Key Encapsulation Mechanismd (KEM). This trait takes slices as arguments.

Required Methods

fn keygen(ek: &mut [u8], dk: &mut [U8], rand: &[U8]) -> Result<(), KeyGenError>

Generate a pair of encapsulation and decapsulation keys. It is the responsibility of the caller to ensure that the rand argument is actually random.

fn encaps( ct: &mut [u8], ss: &mut [U8], ek: &[u8], rand: &[U8], ) -> Result<(), EncapsError>

Encapsulate a shared secret towards a given encapsulation key. It is the responsibility of the caller to ensure that the rand argument is actually random.

fn decaps(ss: &mut [U8], ct: &[u8], dk: &[U8]) -> Result<(), DecapsError>

Decapsulate a shared secret.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.