libcrux_traits::kem

Trait KEM

pub trait KEM {
    type Ciphertext;
    type SharedSecret;
    type EncapsulationKey;
    type DecapsulationKey;

    // Required methods
    fn generate_key_pair(
        rng: &mut impl CryptoRng,
    ) -> Result<KeyPair<Self::DecapsulationKey, Self::EncapsulationKey>, KEMError>;
    fn encapsulate(
        ek: &Self::EncapsulationKey,
        rng: &mut impl CryptoRng,
    ) -> Result<(Self::SharedSecret, Self::Ciphertext), KEMError>;
    fn decapsulate(
        dk: &Self::DecapsulationKey,
        ctxt: &Self::Ciphertext,
    ) -> Result<Self::SharedSecret, KEMError>;
}

This trait captures the required interface of a key encapsulation mechanism (KEM).

Required Associated Types

type Ciphertext

The KEM’s ciphertext.

type SharedSecret

The KEM’s shared secret.

type EncapsulationKey

The KEM’s encapsulation key.

type DecapsulationKey

The KEM’s decapsulation key.

Required Methods

fn generate_key_pair( rng: &mut impl CryptoRng, ) -> Result<KeyPair<Self::DecapsulationKey, Self::EncapsulationKey>, KEMError>

Generate a pair of encapsulation and decapsulation keys.

fn encapsulate( ek: &Self::EncapsulationKey, rng: &mut impl CryptoRng, ) -> Result<(Self::SharedSecret, Self::Ciphertext), KEMError>

Encapsulate a shared secret towards a given encapsulation key.

fn decapsulate( dk: &Self::DecapsulationKey, ctxt: &Self::Ciphertext, ) -> Result<Self::SharedSecret, KEMError>

Decapsulate a shared secret.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.