libcrux_traits

Trait Digest

pub trait Digest<const HASH_LEN: usize> {
    // Required methods
    fn hash(digest: &mut [u8], payload: &[u8]);
    fn update(&mut self, payload: &[u8]);
    fn finish(&self, digest: &mut [u8; HASH_LEN]);
    fn reset(&mut self);
}

A Hash algorithm returning hashes of length HASH_LEN.

Required Methods

fn hash(digest: &mut [u8], payload: &[u8])

Writes the digest for the given input byte slice, into digest in immediate mode.

fn update(&mut self, payload: &[u8])

Add the payload to the digest.

fn finish(&self, digest: &mut [u8; HASH_LEN])

Writes the digest into digest.

Note that the digest state can be continued to be used, to extend the digest.

fn reset(&mut self)

Reset the digest state.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.