libcrux_traits

Module aead

Modules

arrayref

This module contains the trait and related errors for an Authenticated Encryption with Associated Data (AEAD) scheme that takes array references as arguments and writes outputs to mutable array references.

consts

owned

This module contains the trait and related errors for an Authenticated Encryption with Associated Data (AEAD) scheme that takes array references as arguments and returns outputs as arrays or vectors.

slice

This module contains the trait and related errors for an Authenticated Encryption with Associated Data (AEAD) scheme that takes slices as arguments and writes outputs to mutable slices.

tests

typed_owned

This module contains the trait and related errors for an Authenticated Encryption with Associated Data (AEAD) scheme that takes array references as arguments and writes outputs to mutable array references.

typed_refs

This module contains the trait and related errors for an Authenticated Encryption with Associated Data (AEAD) scheme that takes array references as arguments and writes outputs to mutable array references.