libcrux_traits::aead::typed_owned

Trait Aead

pub trait Aead: Sized + AeadConsts {
    type Key;
    type Tag;
    type Nonce;
    type Rand;

    // Required methods
    fn keygen(key: &mut Key<Self>, rand: &Self::Rand) -> Result<(), KeyGenError>;
    fn encrypt(
        ciphertext: &mut [u8],
        tag: &mut Tag<Self>,
        key: &Key<Self>,
        nonce: &Nonce<Self>,
        aad: &[u8],
        plaintext: &[U8],
    ) -> Result<(), EncryptError>;
    fn decrypt(
        plaintext: &mut [U8],
        key: &Key<Self>,
        nonce: &Nonce<Self>,
        aad: &[u8],
        ciphertext: &[u8],
        tag: &Tag<Self>,
    ) -> Result<(), DecryptError>;
}

An Authenticated Encryption with Associated Data (AEAD) scheme. This trait is low-level and is mostly used for implementing other, more usable APIs.

Some implementors of this trait may impose stronger restrictions on the inputs than described here. Check the documentation of the types implementing this trait to make sure which inputs are valid.

Required Associated Types

type Key

The type containing the key data. Usually an array.

type Tag

The type containing the tag data. Usually an array.

type Nonce

The type containing the nonce data. Usually an array.

type Rand

The type containing the randomness to generate a new key. Usually an array.

Required Methods

fn keygen(key: &mut Key<Self>, rand: &Self::Rand) -> Result<(), KeyGenError>

Generate a new key. Consumes the entire randomnes.

fn encrypt( ciphertext: &mut [u8], tag: &mut Tag<Self>, key: &Key<Self>, nonce: &Nonce<Self>, aad: &[u8], plaintext: &[U8], ) -> Result<(), EncryptError>

Encrypt a plaintext message, producing a ciphertext and an authentication tag. The arguments plaintext and ciphertext must have the same length.

fn decrypt( plaintext: &mut [U8], key: &Key<Self>, nonce: &Nonce<Self>, aad: &[u8], ciphertext: &[u8], tag: &Tag<Self>, ) -> Result<(), DecryptError>

Decrypt a ciphertext, verifying its authenticity. The arguments plaintext and ciphertext must have the same length.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.