libcrux_traits::aead::typed_refs

Trait Aead

pub trait Aead: Copy + PartialEq {
    // Required methods
    fn key_len(&self) -> usize;
    fn tag_len(&self) -> usize;
    fn nonce_len(&self) -> usize;
    fn keygen<'a>(
        &self,
        key: KeyMut<'a, Self>,
        rand: &[U8],
    ) -> Result<(), KeyGenError>;
    fn encrypt<'a>(
        &self,
        ciphertext: &mut [u8],
        tag: TagMut<'a, Self>,
        key: KeyRef<'a, Self>,
        nonce: NonceRef<'a, Self>,
        aad: &[u8],
        plaintext: &[U8],
    ) -> Result<(), EncryptError>;
    fn decrypt<'a>(
        &self,
        plaintext: &mut [U8],
        key: KeyRef<'a, Self>,
        nonce: NonceRef<'a, Self>,
        aad: &[u8],
        ciphertext: &[u8],
        tag: TagRef<'a, Self>,
    ) -> Result<(), DecryptError>;

    // Provided methods
    fn new_key<'a>(
        self,
        key: &'a [U8],
    ) -> Result<KeyRef<'a, Self>, WrongLengthError> { ... }
    fn new_key_mut<'a>(
        self,
        key: &'a mut [U8],
    ) -> Result<KeyMut<'a, Self>, WrongLengthError> { ... }
    fn new_tag<'a>(
        self,
        tag: &'a [U8],
    ) -> Result<TagRef<'a, Self>, WrongLengthError> { ... }
    fn new_tag_mut<'a>(
        self,
        tag_mut: &'a mut [U8],
    ) -> Result<TagMut<'a, Self>, WrongLengthError> { ... }
    fn new_nonce<'a>(
        self,
        nonce: &'a [U8],
    ) -> Result<NonceRef<'a, Self>, WrongLengthError> { ... }
}

An Authenticated Encryption with Associated Data (AEAD) scheme. This trait is low-level and is mostly used for implementing other, more usable APIs.

Some implementors of this trait may impose stronger restrictions on the inputs than described here. Check the documentation of the types implementing this trait to make sure which inputs are valid.

Required Methods

fn key_len(&self) -> usize

Returns the length of keys for this algorithm in bytes.

fn tag_len(&self) -> usize

Returns the length of authentication tags for this algorithm in bytes.

fn nonce_len(&self) -> usize

Returns the length of nonces for this algorithm in bytes.

fn keygen<'a>( &self, key: KeyMut<'a, Self>, rand: &[U8], ) -> Result<(), KeyGenError>

Generate a new key for this algorithm using the provided randomness.

fn encrypt<'a>( &self, ciphertext: &mut [u8], tag: TagMut<'a, Self>, key: KeyRef<'a, Self>, nonce: NonceRef<'a, Self>, aad: &[u8], plaintext: &[U8], ) -> Result<(), EncryptError>

Encrypt a plaintext message, producing a ciphertext and an authentication tag. The arguments plaintext and ciphertext must have the same length.

fn decrypt<'a>( &self, plaintext: &mut [U8], key: KeyRef<'a, Self>, nonce: NonceRef<'a, Self>, aad: &[u8], ciphertext: &[u8], tag: TagRef<'a, Self>, ) -> Result<(), DecryptError>

Decrypt a ciphertext, verifying its authenticity. The arguments plaintext and ciphertext must have the same length.

Provided Methods

fn new_key<'a>( self, key: &'a [U8], ) -> Result<KeyRef<'a, Self>, WrongLengthError>

Creates a new key given the algorithm.

fn new_key_mut<'a>( self, key: &'a mut [U8], ) -> Result<KeyMut<'a, Self>, WrongLengthError>

Creates a new mutable key given the algorithm.

fn new_tag<'a>( self, tag: &'a [U8], ) -> Result<TagRef<'a, Self>, WrongLengthError>

Creates a new tag given the algorithm.

fn new_tag_mut<'a>( self, tag_mut: &'a mut [U8], ) -> Result<TagMut<'a, Self>, WrongLengthError>

Creates a new mutable tag given the algorithm.

fn new_nonce<'a>( self, nonce: &'a [U8], ) -> Result<NonceRef<'a, Self>, WrongLengthError>

Creates a new nonce given the algorithm.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.