libcrux_traits::aead::typed_refs

Trait Multiplexes

pub trait Multiplexes<Algo>: Aead + Sized {
    // Required methods
    fn mux_algo(&self) -> Option<Algo>;
    fn wrap_algo(algo: Algo) -> Self;

    // Provided methods
    fn mux_key<'a>(key: KeyRef<'a, Self>) -> Option<KeyRef<'a, Algo>> { ... }
    fn wrap_key<'a>(k: KeyRef<'a, Algo>) -> KeyRef<'a, Self> { ... }
    fn mux_key_mut<'a>(key: KeyMut<'a, Self>) -> Option<KeyMut<'a, Algo>> { ... }
    fn wrap_key_mut<'a>(k: KeyMut<'a, Algo>) -> KeyMut<'a, Self> { ... }
    fn mux_tag<'a>(tag: TagRef<'a, Self>) -> Option<TagRef<'a, Algo>> { ... }
    fn wrap_tag<'a>(tag: TagRef<'a, Algo>) -> TagRef<'a, Self> { ... }
    fn mux_tag_mut<'a>(tag: TagMut<'a, Self>) -> Option<TagMut<'a, Algo>> { ... }
    fn wrap_tag_mut<'a>(tag: TagMut<'a, Algo>) -> TagMut<'a, Self> { ... }
    fn mux_nonce<'a>(nonce: NonceRef<'a, Self>) -> Option<NonceRef<'a, Algo>> { ... }
    fn wrap_nonce<'a>(nonce: NonceRef<'a, Algo>) -> NonceRef<'a, Self> { ... }
}

This trait is implemented by the multiplexing enum, once per actual algorithm. It allows users (and internal code) to convert between structs with the actual algorithm and the multiplexing enum.

Required Methods

fn mux_algo(&self) -> Option<Algo>

If self is actually algorithm Algo, return that.

fn wrap_algo(algo: Algo) -> Self

Create a new multiplexed algorithm value that has algo as actual algorithm.

Provided Methods

fn mux_key<'a>(key: KeyRef<'a, Self>) -> Option<KeyRef<'a, Algo>>

Tries unwrapping the algorithm in a key

fn wrap_key<'a>(k: KeyRef<'a, Algo>) -> KeyRef<'a, Self>

Wraps the algorithm in a key

fn mux_key_mut<'a>(key: KeyMut<'a, Self>) -> Option<KeyMut<'a, Algo>>

Tries unwrapping the algorithm in a mutable key

fn wrap_key_mut<'a>(k: KeyMut<'a, Algo>) -> KeyMut<'a, Self>

Wraps the algorithm in a mutable key

fn mux_tag<'a>(tag: TagRef<'a, Self>) -> Option<TagRef<'a, Algo>>

Tries unwrapping the algorithm in a tag

fn wrap_tag<'a>(tag: TagRef<'a, Algo>) -> TagRef<'a, Self>

Wraps the algorithm in a tag

fn mux_tag_mut<'a>(tag: TagMut<'a, Self>) -> Option<TagMut<'a, Algo>>

Tries unwrapping the algorithm in a mutable tag

fn wrap_tag_mut<'a>(tag: TagMut<'a, Algo>) -> TagMut<'a, Self>

Wraps the algorithm in a mutable tag

fn mux_nonce<'a>(nonce: NonceRef<'a, Self>) -> Option<NonceRef<'a, Algo>>

Tries unwrapping the algorithm in a nonce

fn wrap_nonce<'a>(nonce: NonceRef<'a, Algo>) -> NonceRef<'a, Self>

Wraps the algorithm in a nonce

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.